# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "FirmWizard API",
        "description": "AI-powered marketing platform for law firms",
        "contact": {
            "email": "support@firmwizard.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/crawler/page": {
            "post": {
                "tags": [
                    "Crawler"
                ],
                "summary": "Crawl and store a page for competitor analysis",
                "operationId": "20703fc1044fd465dae49506099a5d0c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "competitor_id",
                                    "url"
                                ],
                                "properties": {
                                    "competitor_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "url": {
                                        "type": "string",
                                        "example": "https://example.com"
                                    },
                                    "save_to_db": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Page crawled successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "page_id": {
                                            "type": "integer"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/crawler/pages/{competitorId}": {
            "get": {
                "tags": [
                    "Crawler"
                ],
                "summary": "Get crawled pages for a competitor",
                "operationId": "77680226ca67857e31822fb545a44c5b",
                "parameters": [
                    {
                        "name": "competitorId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of crawled pages"
                    }
                }
            }
        },
        "/api/crawler/find-competitors": {
            "post": {
                "tags": [
                    "Crawler"
                ],
                "summary": "Find competitors using Google Places API",
                "operationId": "4745db9af1d3f6f8dce358b6c16c7d88",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "collection_id"
                                ],
                                "properties": {
                                    "collection_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Competitor search completed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "run_id": {
                                            "type": "integer"
                                        },
                                        "competitors_found": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/crawler/collections": {
            "get": {
                "tags": [
                    "Crawler"
                ],
                "summary": "Get all collections",
                "operationId": "30bff1bc3dd5ec0ed2cc732d5800581b",
                "responses": {
                    "200": {
                        "description": "List of collections"
                    }
                }
            },
            "post": {
                "tags": [
                    "Crawler"
                ],
                "summary": "Create a new collection",
                "operationId": "021aae38ba3a1f94abfaab62fafcd550",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "firm_id",
                                    "firm_location_id",
                                    "name",
                                    "query"
                                ],
                                "properties": {
                                    "firm_id": {
                                        "type": "integer"
                                    },
                                    "firm_location_id": {
                                        "type": "integer"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "query": {
                                        "type": "string"
                                    },
                                    "enabled": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Collection created"
                    }
                }
            }
        },
...
        "/api/crawler/firm-locations": {
            "get": {
                "tags": [
                    "Crawler"
                ],
                "summary": "Get list of firm locations",
                "operationId": "4c6ae1950477beabe8ec46289dc56404",
                "parameters": [
                    {
                        "name": "firm_id",
                        "in": "query",
                        "description": "Filter by firm ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of firm locations"
                    }
                }
            }
        },
        "/api/crawler/firm-profiles": {
            "get": {
                "tags": [
                    "Crawler"
                ],
                "summary": "Get list of firm profiles",
                "operationId": "5602c500f75ab77c1eefc274d4b6865e",
                "parameters": [
                    {
                        "name": "firm_id",
                        "in": "query",
                        "description": "Filter by firm ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of firm profiles"
                    }
                }
            }
        },
...
    },
...
}